requirements.txt
numpy
scipy
pyserial
matplotlib
PyQt5